+++ libraries/ruby.rb
@@ -10,18 +10,20 @@ module PkgBuild
         name.compact.join('-')
       end
 
-      def gem_name(node, name, ruby_version=nil)
+      def gem_name(node, name, ruby_version=nil, include_rubygem=true)
         gname = []
         gname << node[:pkg_build][:pkg_prefix]
-        gname << 'rubygem'
-        gname << ruby_version if ruby_version
+        gname << 'rubygem' if include_rubygem
+        gname << ruby_version if ruby_version && node[:pkg_build][:ruby][:suffix_version]
         gname << name
         gname.compact.join('-')
       end
 
       def ruby_build(node, version, patchlevel)
-        patch = patchlevel.to_s.start_with?('p') ? patchlevel : "p#{patchlevel}"
-        "#{ruby_name(node, version)}-#{version}-#{patch}"
+        if (patchlevel)
+          patch = patchlevel.to_s.start_with?('p') ? patchlevel : "p#{patchlevel}"
+        end
+        [ruby_name(node, version), version, patch].compact.join('-')
       end
     end
   end