+++ recipes/deps.rb
@@ -1,5 +1,8 @@
 # These are all the things we want all the time. So load them!
 
+Mixlib::ShellOut.send(:remove_const, :DEFAULT_READ_TIMEOUT)
+Mixlib::ShellOut.send(:const_set, :DEFAULT_READ_TIMEOUT, 3600)
+
 include_recipe 'apt'
 include_recipe 'builder'
 include_recipe 'fpm-tng'