+++ recipes/ruby.rb
@@ -9,10 +9,10 @@ comparable_versions = []
 
 versions.uniq.each do |r_ver|
   version, patchlevel = r_ver.split('-')
-  comparable_versions << [Gem::Version.new(version), patchlevel[1,patchlevel.length].to_i]
-  
+  comparable_versions << [Gem::Version.new(version), patchlevel.nil? ? nil : patchlevel.to_s[1,patchlevel.length].to_i]
+
   if(node[:pkg_build][:isolate])
-    pkg_build_isolate "ruby-#{version}-#{patchlevel}" do
+    pkg_build_isolate ['ruby', version, patchlevel].compact.join('-') do
       container 'ubuntu_1204'
       attributes(
         :pkg_build => {
@@ -20,7 +20,8 @@ versions.uniq.each do |r_ver|
             :version => version,
             :patchlevel => patchlevel,
             :suffix_version => node[:pkg_build][:ruby][:suffix_version],
-            :replace_deprecated => node[:pkg_build][:ruby][:replace_deprecated]
+            :replace_deprecated => node[:pkg_build][:ruby][:replace_deprecated],
+            :extra_configure_args => node[:pkg_build][:ruby][:extra_configure_args],
           }
         }
       )
@@ -55,15 +56,21 @@ if(node[:pkg_build][:isolate])
     end.last
     ruby_build = PkgBuild::Ruby.ruby_build(node, install_version.first.version, install_version.last)
     ruby_dpkg = File.join(node[:fpm_tng][:package_dir], "#{ruby_build}.deb")
+
+    commands = ["dpkg -i #{ruby_dpkg}; apt-get -f -q -y install", "gem install --no-ri --no-rdoc rubygems-update"]
+    if(node[:pkg_build][:ruby][:rubygems][:version] == 'latest')
+      commands << "gem update --system"
+    else
+      commands << "gem update --system #{node[:pkg_build][:ruby][:rubygems][:version]}"
+    end
+    commands << "gem install --no-ri --no-rdoc fpm"
+
     node[:pkg_build][:isolated_containers].each do |name, opts|
       lxc_container "#{name}-ruby#{version.version}" do
         action :create
         clone name
         default_fstab false
-        initialize_commands [
-          "dpkg -i #{ruby_dpkg}; apt-get -f -q -y install",
-          'gem install --no-ri --no-rdoc fpm'
-        ]
+        initialize_commands commands
       end
     end
   end