+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.2.4
+* Permit no patch version rubies
+
+## v0.2.3
+* Check for info[:versions] first
+
+## v0.2.2
+* Update passenger build to account for updated module path
+
 ## v0.2.0
 * Allow packages to be built in isolation
 * Tailor support towards `repository` cookbook support (reprepro support still remains)